for #2084, refactor SelectItemExtractor's package
